backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.1
gunicorn==21.2.0
google-cloud-storage==2.10.0
google-cloud-tasks==2.13.2
librosa==0.10.0
essentia==2.1
torch==2.0.1
torchaudio==2.0.2
fb-demucs==1.1.5
spleeter==2.3.2
numpy==1.24.3
scipy==1.11.2
soundfile==0.12.1
